requirements.txt
pypdf2
langchain
google-generativeai
python-dotenv
speechRecognition
pyttsx3
stomp.py==4.1.19
urllib3
logging
langchain_community
langchain_google_genai
faiss-cpu
pyaudio
  